feat: init AccountConnectButton
