+++ templates/buy-me-a-coffee-app/src/onchain/components/index.tsx
@@ -1,0 +1,1 @@
+export * from './accountConnectButton';
